+++ formula-android/src/test/java/com/instacart/formula/android/FlowFactoryTest.kt
@@ -2,7 +2,6 @@ package com.instacart.formula.android
 
 import com.google.common.truth.Truth
 import com.instacart.formula.fragment.FragmentFlowStore
-import com.instacart.formula.integration.DisposableScope
 import com.instacart.formula.android.fakes.TestAccountFragmentContract
 import com.instacart.formula.android.fakes.TestLoginFragmentContract
 import com.instacart.formula.android.fakes.TestSignUpFragmentContract
