+++ formula-android/src/test/java/com/instacart/formula/android/fakes/FakeAuthFlowFactory.kt
@@ -3,7 +3,7 @@ package com.instacart.formula.android.fakes
 import com.instacart.formula.android.Flow
 import com.instacart.formula.android.FlowFactory
 import com.instacart.formula.fragment.FragmentContract
-import com.instacart.formula.integration.DisposableScope
+import com.instacart.formula.android.DisposableScope
 import io.reactivex.rxjava3.core.Observable
 
 class FakeAuthFlowFactory : FlowFactory<FakeComponent, FakeAuthFlowFactory.Component> {
